backend/requirements.txt
# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database & ORM
sqlalchemy==2.0.36
psycopg2-binary==2.9.10

# Configuration
python-dotenv==1.0.1
pydantic==2.10.3

# Security - Password Hashing
bcrypt==4.2.1
argon2-cffi==23.1.0

# Security - JWT Tokens
python-jose[cryptography]==3.3.0

# Security - TOTP
pyotp==2.9.0

# HTTP Client
requests==2.32.3

# Testing
pytest==8.3.4
httpx==0.28.1